# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Social Media App API",
        "description": "REST API for managing social media posts",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "summary": "Get All Posts",
                "description": "Get all posts.",
                "operationId": "get_all_posts_api_posts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PostResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Posts Api Posts Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Post",
                "description": "Create a new post.",
                "operationId": "create_post_api_posts_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/search": {
            "get": {
                "summary": "Search Posts",
                "description": "Search for posts by user or text.",
                "operationId": "search_posts_api_posts_search_get",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search for posts by username",
                            "title": "User"
                        },
                        "description": "Search for posts by username"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search for posts containing this text",
                            "title": "Text"
                        },
                        "description": "Search for posts containing this text"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResponse"
                                    },
                                    "title": "Response Search Posts Api Posts Search Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{post_id}": {
            "get": {
                "summary": "Get Post",
                "description": "Get a specific post by ID.",
                "operationId": "get_post_api_posts__post_id__get",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "summary": "Health Check",
                "description": "Health check endpoint.",
                "operationId": "health_check_api_health_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "summary": "Index",
                "description": "Serve the main HTML page.",
                "operationId": "index__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/style.css": {
            "get": {
                "summary": "Serve Css",
                "description": "Serve CSS file.",
                "operationId": "serve_css_style_css_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/app.js": {
            "get": {
                "summary": "Serve Js",
                "description": "Serve JavaScript file.",
                "operationId": "serve_js_app_js_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HealthResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "HealthResponse"
            },
            "PostCreate": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "user": {
                        "type": "string",
                        "title": "User"
                    }
                },
                "type": "object",
                "required": [
                    "image",
                    "text",
                    "user"
                ],
                "title": "PostCreate"
            },
            "PostResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "user": {
                        "type": "string",
                        "title": "User"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "image",
                    "text",
                    "user",
                    "created_at"
                ],
                "title": "PostResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}